requirements.txt
flask
flask_sqlalchemy
pyodbc
PyJWT
werkzeug
python-dotenv
requests
